refactor: modify relationships between lesson related entities
