+++ src/main/java/UMC/campusNote/mapping/UserLesson.java
@@ -2,7 +2,7 @@ package UMC.campusNote.mapping;
 
 import UMC.campusNote.classSideNote.ClassSideNote;
 import UMC.campusNote.common.BaseEntity;
-import UMC.campusNote.lesson.Lesson;
+import UMC.campusNote.lesson.entity.Lesson;
 import UMC.campusNote.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
